# backend/swagger.yaml
openapi: 3.0.3
info:
  title: ERP System Backend API
  description: |
    Complete API documentation for the ERP System Backend microservices.
    
    ## Services
    - **Auth Service**: Authentication and user management
    - **Organization Service**: Organizations, companies, and locations
    - **Inventory Service**: Products and stock management
    - **License Service**: Application licenses and assignments
    - **Subscription Service**: Subscription plans and billing
    
    ## Authentication
    Most endpoints require JWT authentication via Bearer token in the Authorization header.
    
  version: 1.0.0
  contact:
    name: ERP System Team
    email: support@erpsystem.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081/api/v1
    description: Auth Service (Development)
  - url: http://localhost:8082/api/v1
    description: Organization Service (Development)
  - url: http://localhost:8083/api/v1
    description: Inventory Service (Development)
  - url: http://localhost:8084/api/v1
    description: License Service (Development)
  - url: http://localhost:8085/api/v1
    description: Subscription Service (Development)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or register endpoint
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 200
        message:
          type: string
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        code:
          type: integer
          example: 400
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
    PaginationResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        code:
          type: integer
          example: 200
        data:
          type: array
          items:
            type: object
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            total_pages:
              type: integer
    BaseModel:
      type: object
      properties:
        id:
          type: string
          format: objectid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        created_by:
          type: string
          format: objectid
        updated_by:
          type: string
          format: objectid
          nullable: true
        deleted_at:
          type: string
          format: date-time
          nullable: true
        deleted_by:
          type: string
          format: objectid
          nullable: true
    Address:
      type: object
      required:
        - street
        - city
        - state
        - postal_code
        - country
      properties:
        street:
          type: string
        street2:
          type: string
        city:
          type: string
        state:
          type: string
        postal_code:
          type: string
        country:
          type: string
        country_code:
          type: string
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
    
    # Auth Schemas
...
    Organization:
      allOf:
        - $ref: '#/components/schemas/BaseModel'
        - type: object
          properties:
            name:
              type: string
            legal_name:
              type: string
            domain:
              type: string
            email:
              type: string
              format: email
            phone:
              type: string
            website:
              type: string
            tax_id:
              type: string
            registration_number:
              type: string
            industry:
              type: string
            company_size:
              type: string
              enum: [small, medium, large, enterprise]
            status:
              type: string
              enum: [active, inactive, suspended, trial, cancelled]
            is_active:
              type: boolean
            billing_email:
              type: string
              format: email
            max_users:
              type: integer
            max_companies:
              type: integer
            max_locations:
              type: integer
            current_users:
              type: integer
            current_companies:
              type: integer
            current_locations:
              type: integer
            storage_used_gb:
              type: number
              format: float
            storage_limit_gb:
              type: number
              format: float
    CreateOrganizationRequest:
      type: object
      required:
        - name
        - legal_name
        - domain
        - email
        - billing_email
      properties:
        name:
          type: string
        legal_name:
          type: string
        domain:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        website:
          type: string
        tax_id:
          type: string
        registration_number:
          type: string
        industry:
          type: string
        company_size:
          type: string
          enum: [small, medium, large, enterprise]
        billing_email:
          type: string
          format: email
        max_users:
          type: integer
        max_companies:
          type: integer
        max_locations:
          type: integer
        storage_limit_gb:
          type: number
          format: float
    UpdateOrganizationRequest:
      type: object
      properties:
        name:
          type: string
        legal_name:
          type: string
        domain:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        website:
          type: string
        logo:
          type: string
        is_active:
          type: boolean
    
    # Company Schemas
    Company:
      allOf:
        - $ref: '#/components/schemas/BaseModel'
        - type: object
          properties:
            organization_id:
              type: string
              format: objectid
            name:
              type: string
            legal_name:
              type: string
            code:
              type: string
            tax_id:
              type: string
            registration_number:
              type: string
            vat_number:
              type: string
            email:
              type: string
              format: email
            phone:
              type: string
            address:
              $ref: '#/components/schemas/Address'
            is_active:
              type: boolean
            is_default:
              type: boolean
    CreateCompanyRequest:
      type: object
      required:
        - name
        - legal_name
        - code
        - email
        - address
      properties:
        name:
          type: string
        legal_name:
          type: string
        code:
          type: string
        tax_id:
          type: string
        registration_number:
          type: string
        vat_number:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        is_default:
          type: boolean
    
    # Location Schemas
    Location:
      allOf:
        - $ref: '#/components/schemas/BaseModel'
        - type: object
          properties:
            company_id:
              type: string
              format: objectid
            organization_id:
              type: string
              format: objectid
            name:
              type: string
            code:
              type: string
            type:
              type: string
              enum: [head_office, branch, warehouse, store, factory, distribution_center]
            email:
              type: string
              format: email
            phone:
              type: string
            address:
              $ref: '#/components/schemas/Address'
            is_active:
              type: boolean
            is_default:
              type: boolean
    CreateLocationRequest:
      type: object
      required:
        - name
        - code
        - type
        - email
        - address
      properties:
        name:
          type: string
        code:
          type: string
        type:
          type: string
          enum: [head_office, branch, warehouse, store, factory, distribution_center]
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        is_default:
          type: boolean
    
    # Product Schemas
...
paths:
...
  /organizations:
    get:
      tags:
        - Organizations
      summary: List organizations
      description: Get paginated list of organizations with optional filtering
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, suspended, trial, cancelled]
      responses:
        '200':
          description: Organizations retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
    
    post:
      tags:
        - Organizations
      summary: Create organization
      description: Create a new organization
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
      responses:
        '201':
          description: Organization created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Organization'
        '400':
          description: Validation error or creation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /organizations/{id}:
    get:
      tags:
        - Organizations
      summary: Get organization
      description: Retrieve organization by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Organization retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Organization'
        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    
    put:
      tags:
        - Organizations
      summary: Update organization
      description: Update existing organization
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrganizationRequest'
      responses:
        '200':
          description: Organization updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Organization'
    
    delete:
      tags:
        - Organizations
      summary: Delete organization
      description: Soft delete an organization
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Organization deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /organizations/{id}/companies:
    get:
      tags:
        - Companies
      summary: List companies
      description: Get paginated list of companies under an organization
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Organization ID
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Companies retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
    
    post:
      tags:
        - Companies
      summary: Create company
      description: Create a new company under an organization
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Organization ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyRequest'
      responses:
        '201':
          description: Company created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Company'
  /companies/{id}:
    get:
      tags:
        - Companies
      summary: Get company
      description: Retrieve company by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Company retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Company'
    
    put:
      tags:
        - Companies
      summary: Update company
      description: Update existing company
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                legal_name:
                  type: string
                email:
                  type: string
                is_active:
                  type: boolean
      responses:
        '200':
          description: Company updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Company'
    
    delete:
      tags:
        - Companies
      summary: Delete company
      description: Soft delete a company
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Company deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
  /companies/{id}/locations:
    get:
      tags:
        - Locations
      summary: List locations
      description: Get paginated list of locations under a company
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Company ID
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Locations retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
    
    post:
      tags:
        - Locations
      summary: Create location
      description: Create a new location under a company
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Company ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocationRequest'
      responses:
        '201':
          description: Location created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Location'
  /locations/{id}:
    get:
      tags:
        - Locations
      summary: Get location
      description: Retrieve location by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Location retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Location'
  
  # ===============================
  # INVENTORY SERVICE ENDPOINTS
  # ===============================
...